+++ packages/create/src/cli/index.ts
@@ -10,6 +10,8 @@ import { getRemoveCommands } from './remove/index.js';
 import { getGenerateCommands } from './generate/index.js';
 import { getChangeCommands } from './change/index.js';
 import { getConvertCommands } from './convert/index.js';
+import { getListCommands } from './list/index.js';
+import { getGetCommands } from './get/index.js';
 
 /*
  * We've chosen 'commander' over 'minimist' and 'yargs' for this CLI implementation. Reasons:
@@ -111,6 +113,24 @@ function getCommanderProgram(): Command {
                     The available subcommands are: ${getFeatureSummary(changeCommands.commands)}`
     );
     program.addCommand(changeCommands);
+
+    // Handler for create-fiori list <feature> ..
+    const listCommands = getListCommands();
+    listCommands.description(
+        `Command group for listing saved resources. A subcommand is required.
+                    Usage: \`npx --yes @sap-ux/create@latest list [subcommand] [options]\`
+                    The available subcommands are: ${getFeatureSummary(listCommands.commands)}\n`
+    );
+    program.addCommand(listCommands);
+
+    // Handler for create-fiori get <feature> ..
+    const getCommands = getGetCommands();
+    getCommands.description(
+        `Command group for retrieving saved resources. A subcommand is required.
+                    Usage: \`npx --yes @sap-ux/create@latest get [subcommand] [options]\`
+                    The available subcommands are: ${getFeatureSummary(getCommands.commands)}\n`
+    );
+    program.addCommand(getCommands);
 
     return program;
 }