+++ packages/odata-service-inquirer/src/prompts/datasources/cap-project/questions.ts
@@ -19,6 +19,7 @@ import {
 } from './types.js';
 import { validateCapPath } from './validators.js';
 import { realpath } from 'node:fs/promises';
+import { isAbsolute, resolve } from 'node:path';
 
 /**
  * Find the specified choice in the list of CAP project choices and return its index.
@@ -97,6 +98,12 @@ export function getLocalCapProjectPrompts(
                 }
             },
             guiOptions: { mandatory: true, breadcrumb: t('prompts.capProject.breadcrumb') },
+            filter: (input: string) => {
+                if (PromptState.isYUI) {
+                    return input;
+                }
+                return input && !isAbsolute(input) ? resolve(input) : input;
+            },
             validate: async (projectPath: string): Promise<string | boolean> => {
                 validCapPath = await validateCapPath(projectPath);
                 // Load the cap paths if the path is valid