+++ packages/create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/create
 
+## 0.17.0
+
+### Minor Changes
+
+- f6af206: feat: add CLI commands for managing saved backend systems (#37734)
+
 ## 0.16.8
 
 ### Patch Changes